Adjust for new wgpu
