+++ fastplotlib/utils/__init__.py
@@ -2,7 +2,6 @@ from dataclasses import dataclass
 
 
 from .functions import *
-from ._gpu_info import _notebook_print_banner
 
 
 @dataclass
